# 交付物/接口/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "财务单据智能风险审核系统",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/approval-tasks": {
      "get": {
        "tags": [
          "approval"
        ],
        "summary": "List Approval Tasks",
        "description": "查询当前审批人被分配的任务。",
        "operationId": "list_approval_tasks_api_v1_approval_tasks_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "task_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Task Status"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalTaskPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/approval-tasks/{task_id}": {
      "get": {
        "tags": [
          "approval"
        ],
        "summary": "Get Approval Task",
        "description": "查询当前审批人可见的任务详情。",
        "operationId": "get_approval_task_api_v1_approval_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalTaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/approval-tasks/{task_id}/decision": {
      "post": {
        "tags": [
          "approval"
        ],
        "summary": "Submit Decision",
        "description": "提交审批决定，最终结果不由 AI 自动触发。",
        "operationId": "submit_decision_api_v1_approval_tasks__task_id__decision_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalDecisionCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalDecisionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/approval-history": {
      "get": {
        "tags": [
          "approval"
        ],
        "summary": "Get Approval History",
        "description": "查询单据的固定顺序审批历史。",
        "operationId": "get_approval_history_api_v1_documents__document_id__approval_history_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApprovalHistoryNode"
                  },
                  "title": "Response Get Approval History Api V1 Documents  Document Id  Approval History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/approval-workflows": {
      "get": {
        "tags": [
          "approval"
        ],
        "summary": "List Workflows",
        "description": "管理员查询审批流程配置。",
        "operationId": "list_workflows_api_v1_approval_workflows_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkflowTemplateResponse"
                  },
                  "title": "Response List Workflows Api V1 Approval Workflows Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "approval"
        ],
        "summary": "Create Workflow",
        "description": "管理员创建顺序审批流程草稿。",
        "operationId": "create_workflow_api_v1_approval_workflows_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowCreateCommand"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowTemplateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/approval-workflows/{workflow_id}": {
      "patch": {
        "tags": [
          "approval"
        ],
        "summary": "Update Workflow",
        "description": "管理员更新草稿，已发布模板必须先停用或新建版本。",
        "operationId": "update_workflow_api_v1_approval_workflows__workflow_id__patch",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workflow Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowPatchCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowTemplateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/approval-workflows/{workflow_id}/publish": {
      "post": {
        "tags": [
          "approval"
        ],
        "summary": "Publish Workflow",
        "description": "管理员发布流程，发布原因写入后续审计链路。",
        "operationId": "publish_workflow_api_v1_approval_workflows__workflow_id__publish_post",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workflow Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowPublishCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowTemplateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApprovalDecisionCommand": {
        "properties": {
          "approver_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approver Id"
          },
          "decision": {
            "$ref": "#/components/schemas/DecisionCode"
          },
          "comment": {
            "type": "string",
            "maxLength": 2000,
            "title": "Comment",
            "default": ""
          },
          "review_comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Comment"
          },
          "idempotency_key": {
            "type": "string",
            "maxLength": 128,
            "title": "Idempotency Key",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "ApprovalDecisionCommand",
        "description": "审批人提交决定。"
      },
...
      "ApprovalDecisionResponse": {
        "properties": {
          "task_id": {
            "type": "string",
            "format": "uuid",
            "title": "Task Id"
          },
          "document_status": {
            "type": "string",
            "title": "Document Status"
          },
          "record": {
            "$ref": "#/components/schemas/ApprovalDecisionRecord"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "document_status",
          "record"
        ],
        "title": "ApprovalDecisionResponse",
        "description": "审批决定响应。"
      },
      "ApprovalHistoryNode": {
        "properties": {
          "node_order": {
            "type": "integer",
            "title": "Node Order"
          },
          "node_name": {
            "type": "string",
            "title": "Node Name"
          },
          "assignee_name": {
            "type": "string",
            "title": "Assignee Name"
          },
          "task_status": {
            "type": "string",
            "title": "Task Status"
          },
          "decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision"
          },
          "review_comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Comment"
          },
          "processed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processed At"
          }
        },
        "type": "object",
        "required": [
          "node_order",
          "node_name",
          "assignee_name",
          "task_status"
        ],
        "title": "ApprovalHistoryNode",
        "description": "单据审批历史节点。"
      },
      "ApprovalTaskPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ApprovalTaskResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "total"
        ],
        "title": "ApprovalTaskPage",
        "description": "审批任务分页响应。"
      },
      "ApprovalTaskResponse": {
        "properties": {
          "task_id": {
            "type": "string",
            "format": "uuid",
            "title": "Task Id"
          },
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "document_no": {
            "type": "string",
            "title": "Document No"
          },
          "document_type": {
            "type": "string",
            "title": "Document Type"
          },
          "node_id": {
            "type": "string",
            "format": "uuid",
            "title": "Node Id"
          },
          "node_name": {
            "type": "string",
            "title": "Node Name"
          },
          "node_order": {
            "type": "integer",
            "title": "Node Order"
          },
          "assignee_id": {
            "type": "string",
            "format": "uuid",
            "title": "Assignee Id"
          },
          "assignee_name": {
            "type": "string",
            "title": "Assignee Name"
          },
          "task_status": {
            "type": "string",
            "title": "Task Status"
          },
          "decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision"
          },
          "review_comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Comment"
          },
          "total_amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "overall_risk_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overall Risk Level"
          },
          "pending_finding_count": {
            "type": "integer",
            "title": "Pending Finding Count",
            "default": 0
          },
          "applicant_name": {
            "type": "string",
            "title": "Applicant Name"
          },
          "applicant_department": {
            "type": "string",
            "title": "Applicant Department"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "processed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processed At"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "document_id",
          "document_no",
          "document_type",
          "node_id",
          "node_name",
          "node_order",
          "assignee_id",
          "assignee_name",
          "task_status",
          "total_amount",
          "currency",
          "applicant_name",
          "applicant_department",
          "created_at"
        ],
        "title": "ApprovalTaskResponse",
        "description": "审批任务列表和详情响应。"
      },
...
      "DecisionCode": {
        "type": "string",
        "enum": [
          "approve",
          "return",
          "reject"
        ],
        "title": "DecisionCode",
        "description": "审批决定。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "WorkflowCreateCommand": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Name"
          },
          "document_type": {
            "type": "string",
            "maxLength": 32,
            "title": "Document Type",
            "default": "expense_reimbursement"
          },
          "match_condition": {
            "type": "string",
            "title": "Match Condition",
            "default": ""
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/WorkflowNodeCommand"
            },
            "type": "array",
            "minItems": 1,
            "title": "Nodes"
          }
        },
        "type": "object",
        "required": [
          "name",
          "nodes"
        ],
        "title": "WorkflowCreateCommand",
        "description": "创建审批流程草稿。"
      },
...
      "WorkflowPatchCommand": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "document_type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Type"
          },
          "match_condition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Condition"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "nodes": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/WorkflowNodeCommand"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nodes"
          }
        },
        "type": "object",
        "title": "WorkflowPatchCommand",
        "description": "更新审批流程草稿。"
      },
      "WorkflowPublishCommand": {
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "WorkflowPublishCommand",
        "description": "发布流程前填写的审核原因。"
      },
      "WorkflowTemplateResponse": {
        "properties": {
          "workflow_id": {
            "type": "string",
            "format": "uuid",
            "title": "Workflow Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "document_type": {
            "type": "string",
            "title": "Document Type"
          },
          "match_condition": {
            "type": "string",
            "title": "Match Condition"
          },
          "approval_mode": {
            "type": "string",
            "title": "Approval Mode",
            "default": "sequential"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/WorkflowNodeResponse"
            },
            "type": "array",
            "title": "Nodes"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "workflow_id",
          "name",
          "version",
          "document_type",
          "match_condition",
          "status",
          "nodes",
          "updated_at"
        ],
        "title": "WorkflowTemplateResponse",
        "description": "版本化审批流程响应。"
      }
    }
  }
}